src/rtsp/ml/norm-image-ap/requirements.txt
boto3
awslambdaric
pillow